+++ sites/funmooc/src/backend/base/urls.py
@@ -3,7 +3,7 @@ API routes exposed by our base app.
 """
 from django.urls import re_path
 
-from .views import redirect_edx_courses
+from .views import redirect_edx_resources
 
 # Support both course OpenEdX routes
 #  http://open-fun.fr/courses/course-v1:acme+00001+session01/about
@@ -12,10 +12,18 @@ COURSE_KEY_PATTERN = (
     r"(course-v1:)?(?P<organization>.+)(\/|\+)(?P<course>.+)(\/|\+)(?P<session>.+)"
 )
 
+#  http://open-fun.fr/universities/acme/
+ORGANIZATION_KEY_PATTERN = r"(?P<organization>[^\/]*)"
+
 urlpatterns = [
     re_path(
         r"courses/{}/about/?$".format(COURSE_KEY_PATTERN),
-        redirect_edx_courses,
+        redirect_edx_resources,
+        name="redirect_edx_courses",
+    ),
+    re_path(
+        r"universities/{}/?$".format(ORGANIZATION_KEY_PATTERN),
+        redirect_edx_resources,
         name="redirect_edx_courses",
     ),
 ]