+++ sites/funmooc/src/backend/base/views.py
@@ -7,10 +7,11 @@ from cms.constants import PUBLISHER_STATE_PENDING
 from richie.apps.core.views.error import error_view_handler
 
 
-def redirect_edx_courses(request, organization, course, session):
+def redirect_edx_resources(request, organization, course=None, session=None):
     """
     The richie site is hosted on the same domain as OpenEdX before.
-    Redirect OpenEdX course urls to the corresponding Richie course urls.
+    Redirect OpenEdX course/organization urls
+    to the corresponding Richie course/organization urls.
     """
 
     def get_redirect_url(**kwargs):
@@ -29,9 +30,10 @@ def redirect_edx_courses(request, organization, course, session):
         return page.get_absolute_url(request.LANGUAGE_CODE)
 
     url = (
-        get_redirect_url(course__code__iexact=course)
+        course is not None
+        and get_redirect_url(course__code__iexact=course)
         or get_redirect_url(organization__code__iexact=organization)
-        or get_redirect_url(reverse_id="search")
+        or get_redirect_url(reverse_id="courses")
     )
 
     if url is None: